DataGrid - extract dataGrid.resizing.tests from dataGrid.tests (#15969)
